feat: support forgejo
Adds support for other VCS providers through the new `vcs` and
`vcs_host` arguments, which are optional for Github (the default).

These are meant to be used for the new Forgejo support, which uses
Github Actions-like syntax, enabling minimal changes to get working. The
major change there is using a `mise` fork to enable passing a Github
token directly since Forgejo replaces that with a Forgejo token, which
breaks downloads.